Refactored MimeTypeRenderer API to work with DataThing instances rather than objects and mime type lists.
Added Facetiser API for decomposing DataType instances into 'facets' or
fields. This is a work in progress - it's not displaying data at the
moment, but it isn't crashing either.

Made file chooser locations persistant using the preferences API. It would
be worth rolling this through in all places it makes sence.


git-svn-id: http://taverna.googlecode.com/svn@357 bf327186-88b3-11dd-a302-d386e5130c1c