(Preliminary) processor for connecting to R using Rserv for running R (and S) scripts.
Inspired by the BeanshellProcessor.

To use this processor, R[1] must be installed locally together with Rserv[2], which is to be running on the default port on localhost.

To use, add a R script from Internal services flip. If everything is installed correctly, a nice rserv.png icon will appear next to the processor. Right click it and "Configure rserv". Write the R script in the "Script" tab. There is no need to press any "Save" button.

In the "Ports" tab, add any inputs for the script. The port input will be available as a local variable from the R script using the port name. The java type of the input must be specified as to be converted to the proper R datatypes. Valid inputs are double[], int[] and String[], which will all automatically get the syntactic type l('text/plain'). Thus everything inputted to R is a list, as R likes to do everything in lists.

REXP is a representation of an R value as used internally by R and supported by Rserv. This type is listed as a valid input type, but not currently supported, as there is yet no way to export REXP from an R script. (A checkbox will be added for this later on)

The output of the R script will be the last statement of the script. As such, it is only possible to return one list of values in this version. In later versions, it could be possible to define several output ports as variables which the R script should assign to.


Security Note:

Rserv provides full R functionality, including system() which allows calling operating system commands. Therefore, always run Rserv in localhost-mode, not as a trusted user (preferably a separate R user). Support for non-localhost and non-anonymous connections will be added in later version of this processor.


git-svn-id: http://taverna.googlecode.com/svn@1553 bf327186-88b3-11dd-a302-d386e5130c1c