Cleaned up some code (removing "this.", etc), Java 5 style fields and syntax.
Uses thread safe(r) getProcessors() all over (except add and remove processor)

All initialization done in constructor, not by field definition 1000 lines down in the code. Constructor heavily depends on new method initialize(), also used by clean() so that clean and constructor for certain does the same thing.

Fields moved to top of class.

clone() throws CloneNotSupportedError if we couldn't be cloned, this should reveal weird bugs that would otherwise appear with the previous "Let's just return this" behaviour.

setOffline disables event notifications while it is doing all its clear() and stuff. Correctly sets it back to the PREVIOUS value afterwards, not neccessarily "true".

destroyProcessor puts out a warning if you try to remove a processor that does not exist.


NotifyThread no longer inner class, because that would give implicit reference to the ScuflModel. Instead, it is constructed with references directly to the lists pendingEvents and listeners.

NotifyThread now has a way to be stopped, by setting t.loop=false - which is done in the destructor of ScuflModel. We are still not sure if this destructor will actually be run in a live Taverna, that will be verified using jProfiler later on. This should fix the thread leak of TAV-127.


Added some basic tests that tests the event system and proper construction. Should also test setOffline as it is quite complex with its trials to go on and off etc. In fact, at some point I managed to get an infinite loop with setOffline, where setOffline went online, the checkbox in the GUI changed to reflect this, and then this triggered another call to setOffline again. Hmm..  does not seem to happen with the code of this check-in, though.


git-svn-id: http://taverna.googlecode.com/svn@2008 bf327186-88b3-11dd-a302-d386e5130c1c