TAV-188 and TAV-190
Implemented the ability to have a profile, which is read on startup. This profile acts as a list of required optional components (required beyond those that are dependencies of Workbench) and are loaded when Raven first initialises. The profile is read from a URL defined in raven.properties as raven.profile. This can be overridden by passing as a java parameter (-Draven.profile=....)

The profile also acts as an SPI filter, allowing different profiles for different disciplines.

Within the bootstrap there is also the process of updating a local store of the profile with the remote version. This is then available and and used should the remote profile become unavailable.

Layout saving and loading has been modified to take account of the profile. A layout can now contain artifacts that lack a version. If this is the case the version of the latest component defined in the profile is used - meaning there is a single point of reference for optional component versions (dependencies of workbench still have there versions defined by the dependencies). Upon saving, any artifacts that exist in the profile do not have there versions stored in the layout file.


git-svn-id: http://taverna.googlecode.com/svn@2173 bf327186-88b3-11dd-a302-d386e5130c1c