TAV-211 and TAV-188
Updates are now possible by the user. On first run the default profile bundled with bootstrap is stored locally and used. If the user asks to check for updates, a version comparison is made between the local and the remote. If they differ and update is possible which updates the local profile (whilst backing up the previous one). On next start, Taverna updates the artifacts and uses the new profile.


git-svn-id: http://taverna.googlecode.com/svn@2181 bf327186-88b3-11dd-a302-d386e5130c1c