Support comments in SPI descriptions (but on separate lines). Fixed RendererSPI definition to have comments on separate lines.
Added FIXME that the SPIRegistry might never die because of circular
references to listener. (Also illustrated by moving the repository
listener out with an implicit reference to the SpiRegistry)

The standard s/this.//g fixes

Skip already seen classloaders as previously in TODO

Catch exceptions close to the error, not at the top level. This means
that we won't stop loading SPIs just because say one of the classes in
an SPI definition can't be loaded.

Still needs proper logging instead of System.*.print* and
ex.printStackTrace() craporama


git-svn-id: http://taverna.googlecode.com/svn@2224 bf327186-88b3-11dd-a302-d386e5130c1c