TAV-217
Part done. Refactoring to bring LocalWorkers inline with the same SPI discovery mechansim as other plugins. Descriptions are now stored in a file localworkers.properties that maps the classname to the category:description (this file is actually identical to the old file used to find LocalWorkers).

Moved some biomoby local workers to the biomoby package together with a SPI and properties file.


git-svn-id: http://taverna.googlecode.com/svn@2236 bf327186-88b3-11dd-a302-d386e5130c1c