TAV-221. synchronized (loaderMap) seems to fix the LinkageError problems. Test now tries to load XScufl / jdom, with a hack to include xerces in the path.
Note that the unit test still fails with Maven, but works within Eclipse and running with JUnit 4 main(). (Try moving init() out of synchronized-block)

Still not quite sure if we have fixed the {{LinkageError: duplicate class definition}} exception as I have not been able to reproduce that in tests, probably because I would need to run the unit tests through Raven, which probably IS possible if we make the correct artifacts.


git-svn-id: http://taverna.googlecode.com/svn@2342 bf327186-88b3-11dd-a302-d386e5130c1c