Refactored menubar into an inner class with makeFile(), makeAdvanced(), etc.
TAV-298 exit workflow added through a new method exit(), a new ExitAction, and exit() is also called on window closed so that when we add checks for "Have you saved" (TAV-297) we can do it in one place.

TAV-299 moves the workflows to a separate "Workflows" menu. An upper threshold for refreshWorkflows() (previously refreshFileMenu()) has been added, at 0.3s)


git-svn-id: http://taverna.googlecode.com/svn@2414 bf327186-88b3-11dd-a302-d386e5130c1c