TAV-297 Save changes on exit implemented.
Workbench keeps track of if workflows have been changed
since their creation or opening, by the help of new 
WorkflowChanges, which listens to ScuflModelEvents and tags 
those workflows as "changed" until SaveWorkflowAction
tags it as "synced" again.

WorkflowChanges also remembers which filename a workflow was
most recently opened or save to/from, and this is used as a 
default in SaveWorkflowAction.

In SaveWorkflowAction, if the workflow has not been saved
before, the title is suggested from the Title in Metadata,
unless it is "Untitled Workflow #234123". If it is untitled, 
that  title will be set to the chosen filename instead.


git-svn-id: http://taverna.googlecode.com/svn@2423 bf327186-88b3-11dd-a302-d386e5130c1c