TAV-305
Stores full layout details on exit, but on startup just reads in the split ratio values and updates them in the current perspective. This prevents errors leading to components permanently vanishing from the layout.

Also fixed the accuracy of the dividerlocation. This is now correct and and the divider location remains constant.


git-svn-id: http://taverna.googlecode.com/svn@2435 bf327186-88b3-11dd-a302-d386e5130c1c