TAV-343 switch to a WorkflowPerspective (if needed) on current workflow change.
When performing operations such as File->Open, Taverna will make sure the a WorkflowPerspective instance is the current perspective.

It will be the perspective with the lowest positionHint that is selected unless the current perspective is a WorkflowPerspective, which is then left alone.  

New method setWorkflowPerspective() can be used by UI operations that modify the current workflow, but want to make sure the user can see such changes. 

WorkbenchPerspectives also got a new static ModelMap instead of several calls to ModelMap.getInstance().


git-svn-id: http://taverna.googlecode.com/svn@2576 bf327186-88b3-11dd-a302-d386e5130c1c