TAV-356
Can now hide/show perspectives. Also decoupled EnactPerspective and moved switching to the Enactor perspective to the WorkflowInstanceListener - this makes it totally optional. Infact, when running a workflow the perspective that is swiched to is the first encountered that contains a WorkflowInstanceContainerFactory.

Currently not added to menu's as still some things to do: persist visibility status between Taverna runs, and what to do when all perspectives are hidden (ideally show a blank panel, otherwise prevent last perspective being hidden).


git-svn-id: http://taverna.googlecode.com/svn@2604 bf327186-88b3-11dd-a302-d386e5130c1c