TAV-362
Done the bulk of the functionality. Checks an online XML document, defined by raven.remoteprofiles, that lists all previous versions, which are then displayed in a dialogue box. User can then select the version they want to swich to. Plugins are checked, and if any are incompatible the user is warned giving them the oppurtunity to bail out. The swich eventually takes place by replacing the local taverna-1.5-profile.xml. User then needs to restart to apply this new profile.

Currently the option isn't added to the menu since the UI side of things needs tidying up, and there is not an avaliable remote profile document.


git-svn-id: http://taverna.googlecode.com/svn@2631 bf327186-88b3-11dd-a302-d386e5130c1c