TAV-369
Improvements to startup speed. The problem was caused by the raven bootstrap URLClassloader being around with URLs to remote hosts. When checking for classes, this was being encountered causing each remote jar to be searched for a class that doesn't exist there.

Changed to only use this URLClassloader if actually looking for a raven class.

Also, a similar problem occurs when searching for META-INF/services/<class> when checking for SPI's. Overridden getResources to prevent the parent being search if it is the raven URLClassloader.


git-svn-id: http://taverna.googlecode.com/svn@2653 bf327186-88b3-11dd-a302-d386e5130c1c