TAV-397
The repository, which was being added to TavernaSPIRegistry, wasn't being initialised correctly. Needs to be got from the classloader to get the instance that is correctly associated with the profile. This now works with executeworkflow.sh

Some further testing is required to make sure it works correctly when called via the API as described on teh wiki.


git-svn-id: http://taverna.googlecode.com/svn@2707 bf327186-88b3-11dd-a302-d386e5130c1c