TAV-18 Architecture split in two.
Taverna-wsdl is the front-end that ends up as a .was and provides the WSDL interface. All it does is to start Raven (by using taverna-bootstrap) and then use taverna-engine/.../Engine as the main class - proxying all calls.

Taverna-engine again is inside Raven, and such can do anything Taverna-classes would like to do, such as calling WorkflowLauncher. The engine has a simple queue and processes workflows in the queue.


git-svn-id: http://taverna.googlecode.com/svn@2717 bf327186-88b3-11dd-a302-d386e5130c1c