TAV-411 Support for dependencies for Beanshell.
A new tab "Dependencies" is added to "Configure" that allows you to specify which local JAR files from $HOME/.taverna/lib/ you want to depend on. This is stored in the SCUFL, and a red warning will show if any of these are missing when loading a workflow.

In addition, one can specify if classes (ie. classloaders) are to be reloaded on each invocation, each processor or stay the same for a whole workflow run.


git-svn-id: http://taverna.googlecode.com/svn@2779 bf327186-88b3-11dd-a302-d386e5130c1c