TAV-438
showOptionDialog doesn't directly support radio buttons (as it expects there to be a button which closes when selected, there is no OK button).

Solution was to add radio buttons to a panel and make OK the only option. After OK is clicked the value of the selected radio button is passed on as the output.


git-svn-id: http://taverna.googlecode.com/svn@2865 bf327186-88b3-11dd-a302-d386e5130c1c