CLOSED - issue TAV-396: WSDL processor problem with user provided namespaces http://www.mygrid.org.uk/jira/browse/TAV-396
Overall fix to handling namespaces when building the SOAP body. Significant refactoring of this area of code into a class hierarchy for Literal and Encoded making the code more manageable to change, and easier to write unit tests against.

elementFormDefault is honoured, and namespaces are not declared for child elements if this is unqualified.


git-svn-id: http://taverna.googlecode.com/svn@2920 bf327186-88b3-11dd-a302-d386e5130c1c