TAV-449 TAV-451 TAV-278
Fixed. Threads were being created by the NotifyThread in ScuflModel, in particular (in this case) with respect to Nested workflows.

Changed WP to not add a listener unless the internal model is explicitly accessed for editing. When finished with the ScuflModel is told to remove any listeners.

Modified WP to always use the same listener (rather than create a new one each time) and added a method to tell it to remove its listener from the internal model. This is called with the processor is destroyed within ScuflModel

ScuflModel checks whether a listener is already contained in the list before adding it. This prevents multiple versions of the same listener being added, and only one being removed.


git-svn-id: http://taverna.googlecode.com/svn@2955 bf327186-88b3-11dd-a302-d386e5130c1c