RESOLVED - issue TAV-491: Nested workflow instances never destroyed http://www.mygrid.org.uk/jira/browse/TAV-491
The workflow instance associated with the nested WorfklowTask has destory called on it once the completion or failure event has been fired. The instance itself isn't destroyed until these events are consumed, since the destroy process itself gets stored on the DISPATCH queue.

Updated javadocs for NestedWorkflow events to highlight that the instance contained in these events must be used within that thread, rather being stored for later, otherwise it may be destroyed before it is used.


git-svn-id: http://taverna.googlecode.com/svn@3017 bf327186-88b3-11dd-a302-d386e5130c1c