TAV-498 - avoid 'null' listeners - this is because access to listeners was not properly synchronized.
This patch makes it so that all modifications of listeners is synchronized on listeners. (And the same for pendingEvents). 

Starting/stopping NotifyThread is now handled by checkNotifyThread() - which synchronizes on listeners.

fireModelEvent() synchronizes on listeners and checks if the notifyThread is null before notify-ing it.

setEventStatus() synchronizes on this before setting isFiringEvents.


git-svn-id: http://taverna.googlecode.com/svn@3034 bf327186-88b3-11dd-a302-d386e5130c1c