CLOSED - issue TAV-539: Default admin user http://www.mygrid.org.uk/jira/browse/TAV-539
Created default page to force admin creation if one does not exist. Uses the filtering mechanism to intercept the next restlet and set it to the createadmin resource until the admin has been created at which point the interception is reverted. The avoids the need for a mapping for this page.


git-svn-id: http://taverna.googlecode.com/svn@3247 bf327186-88b3-11dd-a302-d386e5130c1c