TAV-18 Avoid database lockup/deadlock problems
We don't do all this close() all the time, but we do make sure the transaction is finished with a .rollback() for each run of the loop. The close() is done when the thread is finished.

The jobDAO already supports byStatus() and so the findJobsbyStatus() is not needed.

Cleaned up a bit of the logging. 

DAOFactory is shared and not passed around, it's bascically a singleton.


git-svn-id: http://taverna.googlecode.com/svn@3358 bf327186-88b3-11dd-a302-d386e5130c1c