TAV-18 delete on a job says SUCCESS_NO_CONTENT (and not a redirect, which fails the tests) even if the job didn't exists. (According to REST principles)
The job is removed from the queue is it on, not the default queue. The job is removed from the owner's list of jobs. Both of these checks if the job has a queue or owner.

Note that the datadocuments and workflows have to be deleted separately as they could be referenced from other jobs.


git-svn-id: http://taverna.googlecode.com/svn@3359 bf327186-88b3-11dd-a302-d386e5130c1c