Fixed QueueMonitor behaviour.
Transactions are only rolled back it they are active. Prevents the IllegalStateException being thrown at startup causing the queue to terminate.

Reverted logic for Removing completed jobs so that is actually only removes jobs from the queue, and doesn't try to re-remove previously removed jobs.

Reverted logic in starting jobs, so that jobs are assigned to workers if they are available. Removed mid-method return statements.


git-svn-id: http://taverna.googlecode.com/svn@3371 bf327186-88b3-11dd-a302-d386e5130c1c