OPEN - issue TAV-563: Taverna doesn't work with the latest version of R http://www.mygrid.org.uk/jira/browse/TAV-563
The generated script for fetching the variables from R was:

list(a=a, b=b, )

which no longer works in R 2.5:

> list(a=a, b=b, )
the part of the args list of 'list' being evaluated was:
   (a = a, b = b, )
Error: element 3 is empty


while removing the last comma fixes the issues. No longed need a warning box about 2.5.


git-svn-id: http://taverna.googlecode.com/svn@3437 bf327186-88b3-11dd-a302-d386e5130c1c