Defining MIME types for an Activity input and output ports. Some further refactoring required, but this is a good point to commit as the overall logic and tests are working.
Required refactoring includes making the PortDefinitionBean xmlable (currently the use of MimeType in the lists breaks this).This also needs changing to handle multiple Mimetype annotations.
Also need to move the implementation of Activity*PutPort from the api to the impl decouple the building of these ports from the api to the implementation of the Activity.


git-svn-id: http://taverna.googlecode.com/svn@3595 bf327186-88b3-11dd-a302-d386e5130c1c