Moved ActivityInputPort (and Output) out of maelstrom-api to the implementation, and within AbstractActivity rely on a ActivityPortBuilder to create the correct implementations of these ports independantly of the api.
Changed the port definitions to store the mimetypes as Strings. Not entirely happy with this, but its required to allow it to serialized to XML (either that or make MimeType a true Bean). To avoid MimeTypeImpl being in the api module it means that the mimetypes need to be passed as Strings to the ActivityPortBuilder at which point the implementations creates MimeTypeImpl. I'm sure this could be better, but the non mutable nature of annotations will need revisiting anyhow so will give it more thought between now and then....


git-svn-id: http://taverna.googlecode.com/svn@3596 bf327186-88b3-11dd-a302-d386e5130c1c