Dependency injection - added InvocationContext and propagated change through the code. All Event subclasses now have a context in their constructors and the API to AsynchronousActivityCallback has changed.
Tests have been modified to use a shared implementation of InvocationContext called TestInvocationContext, if we need to change the API to InvocationContext there's only one place in the code we need to mess around with.

The dataflow facade implementation now requires a context in its constructor, it then uses that as the root context for all tokens, fire events etc. I've modified the signature of the constructor method in the Edits interface correspondingly


git-svn-id: http://taverna.googlecode.com/svn@4184 bf327186-88b3-11dd-a302-d386e5130c1c