OPEN - issue T2-189: API support for editing of IterationStrategy http://www.mygrid.org.uk/dev/issues/browse/T2-189
Implements MutableTreeNode and getChildren() is no longer mutable. (or at least it doesn't allow modifications to the internal list of children)

TerminalNode exposed as a part of the API - mainly a kind of tagging interface - but done as an abstract class that now makes sure there is not more than 1 child.


git-svn-id: http://taverna.googlecode.com/svn@5903 bf327186-88b3-11dd-a302-d386e5130c1c