T2-242
Changes to hold configuration settings are stored. Values to match the default value are no longer stored in the config  file, meaning that should this default be dynamic (e.g. location of the installation) this value is not stored.
This has a consequence that if a value is deleted, it should not revert to the default value. So, deleted entries that also have a default definition are stored with a encoded value to represent that it has been deleted.


git-svn-id: http://taverna.googlecode.com/svn@6117 bf327186-88b3-11dd-a302-d386e5130c1c