T2-993 Reduce startup time
Optimized last bits that are possible without rewriting Raven. 

Now the main slowdowns are initialization of SPIs and loading of classes - both still slow due to Raven's massive set of classloaders - OSGI should help with some of this.

Service provider menu is now getting the serviceDescriptionRegistry lazily - this chops of almost a second of GUI startup cost.

DataflowRunsComponent now initializes the database in a separate thread (the thread loading previous runs) - this chops of 5 seconds.

FileManagerImpl loads the DataflowPersistenceHandlerRegistry lazily, this chops of another second.

FileOpenRecentMenuAction deserializes the list of previous runs in a separate thread, this chops of almost a second as it avoids XStream initialization lagging down the Menu creation.

This makes new startup time 19 seconds, which should be not too bad.

git-svn-id: http://taverna.googlecode.com/svn@9356 bf327186-88b3-11dd-a302-d386e5130c1c